Plausible attempt at memoization for Parka, not working yet
It infinite loops somewhere while parsing the math expressions.

It's also clear that the memory usage of this can get explosive as the
parsing progresses. There needs to be a scheme for throwing them away
once the earliest possible branch-back point passes a given character.

Given the tradeoffs, and the decent speed of the un-memoized parser on
typical programming language grammars, it's probably fine without.